refactor(wkt): move `_AnyPackable` to its own file
I think we should keep internal-only types in their own files.
